ssl_scan_sock - Convert hardcoded variable 'SNI' to an argument
This makes it much easier to toggle.
Raise an error if the caller wants to use SNI but the host can't be properly parsed
Also group public function at the bottom; this way it's less hidden.

Thanks to Perspectives user Carl for bringing this up and for help with testing.