notary_common - Do not update previous key's observation time
This can hide a serious bug - if the scanner does not run regularly
and consistently, updating the end time of the previous key
can incorrectly fill in a large block of time
where we may have no proof that the key was seen!

At the very least it would be safer to put a cap on this,
such as only extending the end time until the end of the last day
the previous certificate was seen.
However, that also invents data - the certificate may have been taken down.

The correct thing to do is not to update the previous key
so as not to invent or speculate about it at all.