notary_db - Add test for large spaces between observations
If it has been a long time since the last update, do NOT fill in the observation;
create a new one instead.
This safely handles updates if the scanner has not been run in a long time
or the key has not been seen in a long time.
We still want to allow some space, as scans may not happen with the exact
same order or timing, and it may be more than exactly 24 hours
between observations for a given site and key.