notary_db - Change to use contextmanager for sessions
We'll switch to using a context managar so sessions require a 'with' statement:

e.g. 'with ndb.get_session() as session:'

This is a better (and recommended) SQLAlchemy session usage.
This makes sure session use is obvious and enforced.
Sessions will be closed automatically when they are done being used.
This should prevent session leaks and make them much easier to find,
instead of sprinkling the code with a number of 'close_session()' calls.

Remove the open and close functions so there is only one way to do it.