notary_db - Convert get_all_services() to use a raw db connection
Profiling shows this causes the function to use half as much RAM
and run in one third the time with a large number of services (700,000).

This is a good general speedup, but is particularly useful
when loading a large list of services on a machine with limited resources
(e.g. an ec2 micro instance)