threaded_scanner - implement file logging
One of the biggest issues with the version 2 notaries going offline
was eating up too much disk space with infinitely expanding log files.
Instead of simply redirecting the program output as part of the
admin bash script, we'll control it here in python.

We can use a rotating set of log files with a maximum size
to control behaviour and capture exactly what we want.